Start working on #232
